[NO-ISSUE] 타입으로 이름 세분화
