+++ Projects/Domain/DomainModelsDomain/Interface/ItemList.swift
@@ -1,18 +1,6 @@
 import Foundation
 
-public struct ItemListInfo {
-    public let id: String?
-    public let name: String?
-    
-    public init(id: String?, name: String?) {
-        self.id = id
-        self.name = name
-    }
-}
-
-public enum ItemList {
-    case category(ItemListInfo)
-    case search(ItemListInfo)
-    case wishlist(ItemListInfo)
-    case tbd
+public protocol ItemList {
+    var itemListId: String? { get }
+    var itemListName: String? { get }
 }
