+++ Projects/Domain/ItemListInteractionDomain/Interface/ItemListInteractionRepository.swift
@@ -5,10 +5,10 @@ import DomainModelsDomainInterface
 public protocol ItemListInteractionRepository {
     func sendViewItemList(
         items: [ItemListItem],
-        itemList: ItemList
+        itemList: ItemListType
     )
     func sendSelectItem(
         items: [ItemListItem],
-        itemList: ItemList
+        itemList: ItemListType
     )
 }
